[poly-fitter-minor-1]
1. added `polynomial_f.py`
2. removed `others.py`
3. updated `__init__.py`
